method hiding in java
Method Hiding occurs when a subclass defines a static method with the same signature as a static method in the parent class. Unlike method overriding (where instance methods are involved), in method hiding, the method call is resolved based on the class type at compile time, not at runtime. Therefore, even if an object of a subclass is referred to by a parent class reference, the parent class's version of the static method is invoked.